iommu: arm-smmu: Allow inherting stream mapping from bootloader
The Qualcomm bootloaders leaves the IOMMU with stream mapping for
the display hardware to be able to read the framebuffer memory in DDR,
to continuously display a boot splash or to implement EFI framebuffer.

This patch implements support for implementations to pin stream mappings
and adds the code to the Qualcomm implementation for reading out the
stream mapping from the bootloader, with the result of maintaining the
display hardware's access to DDR until the context bank is enabled.

Without this you have to "fastboot oem select-display-panel none" before
booting to avoid a kernel panic during the SMMU probes; this can be
reverted on the OP5/5T using "fastboot oem select-display-panel prim".

Heavily based on downstream implementation by Patrick Daly
<pdaly@codeaurora.org>.

Link: https://lore.kernel.org/r/20191226221709.3844244-4-bjorn.andersson@linaro.org
From: https://github.com/andersson/kernel/commit/d1b203d8dacf8ebbab2540c343c24506b7bec937
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Jami Kettunen <jami.kettunen@protonmail.com>